refactor : 추상화 통일
