+++ src/app/(main)/profile/setting/edit/components/step1/Step1.client.tsx
@@ -1,0 +1,17 @@
+import Step1Header from './Step1Header';
+import Step1InputForm from './Step1InputForm.client';
+import { Spacing } from '@/components/common/Spacing';
+
+interface Step1Props {
+  onNext: () => void;
+}
+
+export default function Step1({ onNext }: Step1Props) {
+  return (
+    <form>
+      <Step1Header />
+      <Spacing size={20} />
+      <Step1InputForm onNext={onNext} />
+    </form>
+  );
+}
