+++ src/app/(main)/profile/setting/edit/components/step2/Step2.client.tsx
@@ -1,5 +1,5 @@
-import InputForm from './components/InputForm.client';
-import PersonalityEditHeader from './components/PersonalityEditHeader';
+import Step2Header from './Step2Header';
+import Step2InputForm from './Step2InputForm.client';
 import { Spacing } from '@/components/common/Spacing';
 
 interface Step2Props {
@@ -9,7 +9,7 @@ interface Step2Props {
 export default function Step2({ onPrev }: Step2Props) {
   return (
     <>
-      <PersonalityEditHeader onClose={onPrev} />
+      <Step2Header onClose={onPrev} />
       <div className="px-20 pb-16 pt-32 text-h3 text-sign-cto">
         사용자님의 성격을
         <br />
@@ -17,7 +17,7 @@ export default function Step2({ onPrev }: Step2Props) {
       </div>
       <p className="px-20 text-subtitle-2 text-sign-tertiary">3개 이상 선택해주세요.</p>
       <Spacing size={16} />
-      <InputForm onClose={onPrev} />
+      <Step2InputForm onClose={onPrev} />
     </>
   );
 }
